break down conftest into smaller fixture files
